make some code more sharable
