+++ ui/src/components/index.ts
@@ -2,4 +2,5 @@ export * from './ModelUsage';
 export * from './MultiModelDemo';
 export * from './Predict';
 export * from './Interpret';
+export * from './Interpreters';
 export * from './TaskCards';
