+++ ui/src/lib/index.ts
@@ -1,4 +1,3 @@
 export * from './ModelInfo';
-export * from './Saliency';
 export * from './TaskCard';
 export * from './InterpreterId';
