+++ ui/src/tugboat/components/index.ts
@@ -6,6 +6,7 @@ export * from './ModelCard';
 export * from './MultiModelDemo';
 export * from './PrettyPrintedJSON';
 export * from './Promised';
+export * from './Saliency';
 export * from './SelectExample';
 export * from './SelectModel';
 export * from './shared';
